Fixed frustum changes
Fixed the order of the matrix multiplication in the bounding box shader and added a new rejection test.
The ReduceDepth function is now called in every frame for the IZB rendering as well.
Changed the head texture size for now 1024 gives the best performance.
Added the near/far clipping plane change based on the ReduceDepth values.